Add: server and some types
This is just a lil baby prototype. The whole idea is that types are the
language between client and server. they use json to serialize between
em.

There is a big bug in the server, that is because I am lazy and this is
a prototype. Currently the messages are structured like
```
<name>: <params struct serialized to json>
```

it returns the return struct serialized to json.

The packet structure was chosen cause im lazy and its easy. having it be
an actual json rpc kinda *sucks* because we would quickly get into a bit
of type hell and the alternative is using the dynamic feature of the
json parser.

The bug comes from how we get `<name>`, we just read till `:`. This is
super lazy and clearly not a good idea but hey who cares. Essentially we
get stuck reading the json.

How would I improve this?

This currently only supports a single client with little to no error
handling. There is no concurrency at all so I gotta add that.

currently its using a stream so if a client does multiple requests over
the same stream, no way to properly route responses to their callers.

Overall theres a bunch of things to do but these simple things have been
done.